scraper/requirements.txt
# Scrapers
requests>=2.31.0
pyyaml>=6.0
python-slugify>=8.0.0

# Board scout — NOTE: install as 'python-jobspy', NOT 'jobspy' (different package)
python-jobspy>=1.1.0

# API backend
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# Claude API
anthropic>=0.25.0

# Slack
slack-sdk>=3.26.0
